now u can use /g without command

+++ src/main.ts
@@ -6,7 +6,7 @@ import { animation } from './animation'
 import Dictionary from './dictionary'
 import {Game, Word} from './game'
 
-dotenv.config({path:"../.env"})
+dotenv.config()
 
 if(!process.env.BOT_TOKEN){
     console.error("Discord token is not defined!")
@@ -117,6 +117,10 @@ export class GallowsBotInterface{
     })
     
     client.on('message', msg => {
+        //@ts-ignore
+        if(msg.author.id === client.user.id && client) return
+        
+        
         const CurrentUser = new DiscordUser(
             msg.author.id,
             msg.author.username
@@ -135,11 +139,51 @@ export class GallowsBotInterface{
             // if()
             return
         }
+        const guessLogic = (guess: string) => {
+            if(!CurrentState.isGameGoing){ return }
+            GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.GUESS, CurrentUser, CurrentState)
+            try {
+                const word = GameEngine.openLetter(Word.from(CurrentState.word), guess)   
+                UsersManager.setState(msg.author.id, (prev) => ({...prev, word, isGameGoing: !word.isOppened, lifes: CurrentState.lifes + 1 > 7 ? 7 : CurrentState.lifes + 1}))
+                if(word.isOppened){
+                    GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.WON, CurrentUser, CurrentState)
+                    msg.reply("Вы угадали слово это: - "+word.text)
+                    setTimeout(() => {
+                        msg.guild?.channels.cache.get(CurrentState.chanelID)?.delete()
+                        GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.CHANEL_DELETED, CurrentUser, CurrentState)
+                    }, 2000)
+                    return
+                }
+                msg.reply(word.text)             
+            } catch (error) { 
+                if(CurrentState.lifes > 1){
+                    UsersManager.setState(CurrentUser.Discord_id, (prev) => ({...prev, lifes: CurrentState.lifes - 1}))
+                    
+                    msg.reply(`
+${animation[CurrentState.lifes - 1]}
+Этой буквы нету. Осталось жизней! - ${CurrentState.lifes - 1}`) 
+                }else{
+                    GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.LOSE, CurrentUser, CurrentState)
+                    msg.reply('Вы проиграли! Это было слово! "'+CurrentState.word.word+'"')
+                    UsersManager.setState(msg.author.id, (prev) => ({...prev, isGameGoing: false}))
+                    setTimeout(() => {
+                        msg.guild?.channels.cache.get(CurrentState.chanelID)?.delete()
+                        GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.CHANEL_DELETED, CurrentUser, CurrentState)
+                    }, 2000)
+                }
+            }            
+        }
 
+
+        if(msg.channel.id === CurrentState.chanelID){
+            guessLogic(msg.content)
+            return
+        }
+        
+        // if()
         switch(command){
             case 'start':
             msg.delete()
-            GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.START, CurrentUser, CurrentState)
             
             if(CurrentState.isGameGoing === false){
                 const newWord =  GameEngine.createWord(Dictionary.getWord())
@@ -148,55 +192,33 @@ export class GallowsBotInterface{
                     type: 'text',
                     permissionOverwrites: [
                         {
+                            id: msg.guild.roles.everyone,
+                            deny: ["VIEW_CHANNEL", "SEND_MESSAGES", "READ_MESSAGE_HISTORY"]
+                        },
+                        {
                             id: msg.author.id,
                             allow: ["VIEW_CHANNEL", "SEND_MESSAGES", "READ_MESSAGE_HISTORY"]
                         }
-                    ]
+                    ],
+
                 }).then((Chanel) => {
                     GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.CHANEL_CREATED, CurrentUser, CurrentState)
                     Chanel.send("Игра будет происходить в этом канале!")
                     Chanel.send("Длина слова - "+newWord.chars.length)
+                    
                     if(isNewUser){
                         Chanel.send("АБОБА НАПИСАЛ ПРАВИЛА ИГРЫ(ВСЕ В ШОКЕ)!!!")//rules here
                     }
                     UsersManager.setState(msg.author.id, (prev) => ({...prev, chanelID: Chanel.id}))
+                    GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.START, CurrentUser, UsersManager.getState(CurrentUser.Discord_id))
+
                 })
               }else{
                 msg.reply("Вы уже запустили игру!")  //
               }
             break;
             case 'g':
-                GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.GUESS, CurrentUser, CurrentState)
-                try {
-                    const word = GameEngine.openLetter(Word.from(CurrentState.word), args[0])   
-                    UsersManager.setState(msg.author.id, (prev) => ({...prev, word, isGameGoing: !word.isOppened, lifes: CurrentState.lifes + 1 > 7 ? 7 : CurrentState.lifes + 1}))
-                    if(word.isOppened){
-                        GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.WON, CurrentUser, CurrentState)
-                        msg.reply("Вы угадали слово это: - "+word.text)
-                        setTimeout(() => {
-                            msg.guild?.channels.cache.get(CurrentState.chanelID)?.delete()
-                            GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.CHANEL_DELETED, CurrentUser, CurrentState)
-                        }, 2000)
-                        return
-                    }
-                    msg.reply(word.text)             
-                } catch (error) { 
-                    if(CurrentState.lifes > 1){
-                        UsersManager.setState(CurrentUser.Discord_id, (prev) => ({...prev, lifes: CurrentState.lifes - 1}))
-                        
-                        msg.reply(`
-${animation[CurrentState.lifes - 1]}
-Этой буквы нету. Осталось жизней! - ${CurrentState.lifes - 1}`) 
-                    }else{
-                        GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.LOSE, CurrentUser, CurrentState)
-                        msg.reply('Вы проиграли! Это было слово! "'+CurrentState.word.word+'"')
-                        UsersManager.setState(msg.author.id, (prev) => ({...prev, isGameGoing: false}))
-                        setTimeout(() => {
-                            msg.guild?.channels.cache.get(CurrentState.chanelID)?.delete()
-                            GallowsBotInterface.dispatch(GallowsBotInterface.EVENTS.CHANEL_DELETED, CurrentUser, CurrentState)
-                        }, 2000)
-                    }
-                }
+                guessLogic(args[0])
             break;
         }
     });
